Work around return-value issue.
While it's likely _some kind_ of bug that a non-zero value is returned from the test (and also just when only importing the package-under-test and exiting, so it's not the rest of the test), that's not something we test for here (as the package is build correctly). Nor is it that likely to impact much.
However, Conan makes an exception out of it, which caused a crash in the test, even though it passes all of the test-cases correctly and outputs 'True' at the end like it should.

done as part of CURA-11622